Move resolution related classes to the menu package
