+++ engine/src/main/java/org/destinationsol/menu/Resolution.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.destinationsol;
+package org.destinationsol.menu;
 
-class Resolution {
+public class Resolution {
     private final int width;
     private final int height;
